Change the size of blinkScript programs
To save a bit memory the types used by blinkScript programs were changed. All opcodes are just one byte and blinkScript shouldn't exceed 255 objects/constants. Therefore it should be acceptable to cut the memory usage of opcodes in half.

+++ src/firmware/lib/blinkscript/src/printer/disassembler.cpp
@@ -105,31 +105,31 @@ namespace SyncBlink
             {
             case OpCode::VALUE:
                 ss << "VALUE\n";
-                ss << "\t" << std::setw(3) << ++i << ": ." << program.getCode()[i] << "\n";
+                ss << "\t" << std::setw(3) << ++i << ": ." << (int)program.getCode()[i] << "\n";
                 break;
             case OpCode::CLOSURE:
                 ss << "CLOSURE\n";
-                ss << "\t" << std::setw(3) << ++i << ": ." << program.getCode()[i] << "\n";
+                ss << "\t" << std::setw(3) << ++i << ": ." << (int)program.getCode()[i] << "\n";
                 break;
             case OpCode::DEFINE_GLOBAL:
                 ss << "DEFINE_GLOBAL\n";
-                ss << "\t" << std::setw(3) << ++i << ": ." << program.getCode()[i] << "\n";
+                ss << "\t" << std::setw(3) << ++i << ": ." << (int)program.getCode()[i] << "\n";
                 break;
             case OpCode::LOAD_GLOBAL:
                 ss << "LOAD_GLOBAL\n";
-                ss << "\t" << std::setw(3) << ++i << ": ." << program.getCode()[i] << "\n";
+                ss << "\t" << std::setw(3) << ++i << ": ." << (int)program.getCode()[i] << "\n";
                 break;
             case OpCode::LOAD_LOCAL:
                 ss << "LOAD_LOCAL\n";
-                ss << "\t" << std::setw(3) << ++i << ": ." << program.getCode()[i] << "\n";
+                ss << "\t" << std::setw(3) << ++i << ": ." << (int)program.getCode()[i] << "\n";
                 break;
             case OpCode::SET_GLOBAL:
                 ss << "SET_GLOBAL\n";
-                ss << "\t" << std::setw(3) << ++i << ": ." << program.getCode()[i] << "\n";
+                ss << "\t" << std::setw(3) << ++i << ": ." << (int)program.getCode()[i] << "\n";
                 break;
             case OpCode::SET_LOCAL:
                 ss << "SET_LOCAL\n";
-                ss << "\t" << std::setw(3) << ++i << ": ." << program.getCode()[i] << "\n";
+                ss << "\t" << std::setw(3) << ++i << ": ." << (int)program.getCode()[i] << "\n";
                 break;
             case OpCode::SET_INDEX:
                 ss << "SET_INDEX\n";
@@ -196,7 +196,7 @@ namespace SyncBlink
                 break;
             case OpCode::CLEAR_SCOPE:
                 ss << "CLEAR_SCOPE\n";
-                ss << "\t" << std::setw(3) << ++i << ": ." << program.getCode()[i] << "\n";
+                ss << "\t" << std::setw(3) << ++i << ": ." << (int)program.getCode()[i] << "\n";
                 break;
             }
         }
